Replace DOS newline CR/LF with Unix style LF (performed by dos2unix).
While editor is running, files  ctypes.util, mainWindow, os, and wx, one for
each of the imported packages, are created. Because of the line ending of the
script that imported them (editor.py), their files names in Unix systems will
end in '\r' making it really hard to remove (must type Ctrl + M to enter '\r').

This was causing the creation